requirements.txt
# FastAPI and core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Twilio
twilio==8.10.0

# LiveKit
livekit-api==0.4.0
livekit-agents==0.4.0
livekit-plugins-openai==0.4.0

# OpenAI (for LLM)
openai==1.12.0

# HTTP client
httpx==0.26.0

# Environment
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
